+++ demo/RowComponents/index.js
@@ -1,6 +1,8 @@
 import React from 'react';
 
 import TextLabelDoc from './TextLabelDoc';
+import ButtonDoc from './ButtonDoc';
+import IconButtonDoc from './IconButtonDoc';
 
 function RowComponents() {
     return (
@@ -8,6 +10,8 @@ function RowComponents() {
             <h1>Row Components</h1>
 
             <TextLabelDoc />
+            <ButtonDoc />
+            <IconButtonDoc />
         </div>
     );
 }