+++ src/index.js
@@ -6,14 +6,19 @@ import Icon from './Icon';
 import StatusIcon from './StatusIcon';
 import Tag from './Tag';
 import Text from './Text';
+import Tooltip from './Tooltip';
+import AnchoredTooltip from './AnchoredTooltip';
 
 // Layout helpers
 import FlexCell from './FlexCell';
+import IconLayout from './IconLayout';
 import RowCompBody from './RowCompBody';
 import TextEllipsis from './TextEllipsis';
 
 // Row components
 import TextLabel from './TextLabel';
+import Button from './Button';
+import IconButton from './IconButton';
 
 export {
     BasicRow,
@@ -21,10 +26,15 @@ export {
     StatusIcon,
     Tag,
     Text,
+    Tooltip,
+    AnchoredTooltip,
 
     FlexCell,
+    IconLayout,
     RowCompBody,
     TextEllipsis,
 
     TextLabel,
+    Button,
+    IconButton,
 };